third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-misc/timeline-animation-frame.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/timeline-test.js"></script>
<script>
function performActions()
{
    var callback;
    var promise = new Promise((fulfill) => callback = fulfill);
    var requestId = window.requestAnimationFrame(animationFrameCallback);
    function animationFrameCallback()
    {
        window.cancelAnimationFrame(requestId);
        if (callback)
            callback();
    }
    return promise;
}

function test() {
  PerformanceTestRunner.invokeAsyncWithTimeline('performActions', finish);

  function finish() {
    PerformanceTestRunner.printTimelineRecordsWithDetails('RequestAnimationFrame');
    PerformanceTestRunner.printTimelineRecordsWithDetails('FireAnimationFrame');
    PerformanceTestRunner.printTimelineRecordsWithDetails('CancelAnimationFrame');
    TestRunner.completeTest();
  }
}

if (!window.testRunner)
    setTimeout(performActions, 2000);
</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline events for Animation Frame feature
</p>
</body>
</html>
