third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-misc/timeline-animation-frame-expected.txt

Tests the Timeline events for Animation Frame feature

RequestAnimationFrame Properties:
{
    data : {
        frame : <string>
        id : <number>
        stackTrace : <object>
    }
    endTime : <number>
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    type : "RequestAnimationFrame"
}
Text details for RequestAnimationFrame: timeline-animation-frame.html:10
FireAnimationFrame Properties:
{
    data : {
        frame : <string>
        id : <number>
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "FireAnimationFrame"
}
Text details for FireAnimationFrame: timeline-animation-frame.html:10
CancelAnimationFrame Properties:
{
    data : {
        frame : <string>
        id : <number>
        stackTrace : <object>
    }
    endTime : <number>
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    type : "CancelAnimationFrame"
}
Text details for CancelAnimationFrame: timeline-animation-frame.html:13
